[Refactor] #39 - AddToDoPriorityView 구조 개선 및 UI 디테일 수정
