+++ Memento-iOS/Memento-iOS/Source/Presentation/Add/ToDo/View/AddToDoToolbarView.swift
@@ -102,7 +102,7 @@ struct AddToDoToolbarView: View {
         .frame(width: 42, height: 42)
         .sheet(isPresented: $viewModel.isPriorityPickerPresented) {
             PickerSheet(type: .addToDo(.priority)) {
-                AddTodoPriorityView(
+                AddToDoPriorityView(
                     viewModel: viewModel,
                     viewType: .add
                 )
